+++ crates/bevy_editor_pls_core/src/editor_window.rs
@@ -23,6 +23,9 @@ pub trait EditorWindow: 'static {
     fn viewport_toolbar_ui(world: &mut World, cx: EditorWindowContext, ui: &mut egui::Ui) {
         let _ = (world, cx, ui);
     }
+    fn viewport_ui(world: &mut World, cx: EditorWindowContext, ui: &mut egui::Ui) {
+        let _ = (world, cx, ui);
+    }
 
     fn app_setup(app: &mut App) {
         let _ = app;