+++ crates/bevy_editor_pls_default_windows/src/gizmos.rs
@@ -1,0 +1,86 @@
+use bevy::{prelude::*, render::camera::CameraProjection};
+
+use bevy_editor_pls_core::editor_window::{EditorWindow, EditorWindowContext};
+use bevy_inspector_egui::{bevy_inspector::hierarchy::SelectedEntities, egui};
+use egui_gizmo::GizmoMode;
+
+use crate::{
+    cameras::{ActiveEditorCamera, CameraWindow},
+    hierarchy::HierarchyWindow,
+};
+
+pub struct GizmoState {
+    pub camera_gizmo_active: bool,
+    pub gizmo_mode: GizmoMode,
+}
+
+impl Default for GizmoState {
+    fn default() -> Self {
+        Self {
+            camera_gizmo_active: true,
+            gizmo_mode: GizmoMode::Translate,
+        }
+    }
+}
+
+pub struct GizmoWindow;
+
+impl EditorWindow for GizmoWindow {
+    type State = GizmoState;
+
+    const NAME: &'static str = "Gizmos";
+
+    fn ui(_world: &mut World, _cx: EditorWindowContext, ui: &mut egui::Ui) {
+        ui.label("Gizmos can currently not be configured");
+    }
+
+    fn viewport_toolbar_ui(world: &mut World, cx: EditorWindowContext, ui: &mut egui::Ui) {
+        let gizmo_state = cx.state::<GizmoWindow>().unwrap();
+
+        if gizmo_state.camera_gizmo_active {
+            if let (Some(hierarchy_state), Some(_camera_state)) =
+                (cx.state::<HierarchyWindow>(), cx.state::<CameraWindow>())
+            {
+                draw_gizmo(ui, world, &hierarchy_state.selected, gizmo_state.gizmo_mode);
+            }
+        }
+    }
+}
+
+fn draw_gizmo(
+    ui: &mut egui::Ui,
+    world: &mut World,
+    selected_entities: &SelectedEntities,
+    gizmo_mode: GizmoMode,
+) {
+    let (cam_transform, projection) = world
+        .query_filtered::<(&GlobalTransform, &Projection), With<ActiveEditorCamera>>()
+        .single(world);
+    let view_matrix = Mat4::from(cam_transform.affine().inverse());
+    let projection_matrix = projection.get_projection_matrix();
+
+    if selected_entities.len() != 1 {
+        return;
+    }
+
+    for selected in selected_entities.iter() {
+        let Some(transform) = world.get::<Transform>(selected) else { continue };
+        let model_matrix = transform.compute_matrix();
+
+        let Some(result) = egui_gizmo::Gizmo::new(selected)
+                    .model_matrix(model_matrix.to_cols_array_2d())
+                    .view_matrix(view_matrix.to_cols_array_2d())
+                    .projection_matrix(projection_matrix.to_cols_array_2d())
+                    .orientation(egui_gizmo::GizmoOrientation::Local)
+                    .mode(gizmo_mode)
+                    .interact(ui)
+                else { continue };
+
+        let mut transform = world.get_mut::<Transform>(selected).unwrap();
+        *transform = Transform {
+            translation: Vec3::from(<[f32; 3]>::from(result.translation)),
+            rotation: Quat::from_array(<[f32; 4]>::from(result.rotation)),
+            scale: Vec3::from(<[f32; 3]>::from(result.scale)),
+        };
+    }
+}