+++ crates/bevy_editor_pls_default_windows/src/lib.rs
@@ -3,6 +3,7 @@ pub mod assets;
 pub mod cameras;
 pub mod debug_settings;
 pub mod diagnostics;
+pub mod gizmos;
 pub mod hierarchy;
 pub mod inspector;
 pub mod renderer;